--- apps/web/src/components/forms/index.ts
@@ -1,3 +1,0 @@
-export { PollDetailsForm } from "./poll-details-form";
-export { default as PollOptionsForm } from "./poll-options-form/poll-options-form";
-export * from "./types";
