--- apps/web/src/components/forms/poll-options-form/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./poll-options-form";
-export * from "./types";
