--- apps/web/src/components/forms/poll-options-form/month-calendar/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./month-calendar";
