+++ apps/web/src/features/poll/components/create-poll.tsx
@@ -32,10 +32,11 @@ import React from "react";
 import { useForm } from "react-hook-form";
 import useFormPersist from "react-hook-form-persist";
 import { useCopyToClipboard, useUnmount } from "react-use";
-import type { NewEventData } from "@/components/forms";
-import { PollDetailsForm, PollOptionsForm } from "@/components/forms";
-import { PollSettingsForm } from "@/components/forms/poll-settings";
 import { useUser } from "@/components/user-provider";
+import { PollDetailsForm } from "@/features/poll/components/forms/poll-details-form";
+import PollOptionsForm from "@/features/poll/components/forms/poll-options-form/poll-options-form";
+import { PollSettingsForm } from "@/features/poll/components/forms/poll-settings";
+import type { NewEventData } from "@/features/poll/components/forms/types";
 import { Trans, useTranslation } from "@/i18n/client";
 import { trpc } from "@/trpc/client";
